check the file exists before trying to use it

+++ src/Events/Executed.php
@@ -18,16 +18,18 @@ class Executed extends Event
         parent::__construct($task);
 
         $time_elapsed_secs = microtime(true) - $started;
+        
+        if(file_exists(storage_path($task->getMutexName()))) {
+           $output = file_get_contents(storage_path($task->getMutexName()));
 
-        $output = file_get_contents(storage_path($task->getMutexName()));
+            $task->results()->create([
+                'duration'  => $time_elapsed_secs * 1000,
+                'result'    => $output,
+            ]);
 
-        $task->results()->create([
-            'duration'  => $time_elapsed_secs * 1000,
-            'result'    => $output,
-        ]);
+            unlink(storage_path($task->getMutexName()));
 
-        unlink(storage_path($task->getMutexName()));
-
-        $task->notify(new TaskCompleted($output));
+            $task->notify(new TaskCompleted($output)); 
+        }
     }
 }
